Update monitoring layout
