+++ packages/website/src/common/OneColorSwitch/index.tsx
@@ -1,0 +1,20 @@
+import { Switch, withStyles } from '@material-ui/core';
+
+const OneColorSwitch = withStyles((theme) => ({
+  switchBase: {
+    color: theme.palette.primary.main,
+    '&$checked': {
+      color: theme.palette.primary.main,
+    },
+    '&$checked + $track': {
+      backgroundColor: theme.palette.primary.main,
+    },
+    '& + $track': {
+      backgroundColor: theme.palette.primary.main,
+    },
+  },
+  checked: {},
+  track: {},
+}))(Switch);
+
+export default OneColorSwitch;
