+++ packages/website/src/routes/MonitoringRoutes/SiteMetrics/index.test.tsx
@@ -1,0 +1,54 @@
+import React from 'react';
+import { Provider as ReduxProvider } from 'react-redux';
+import { BrowserRouter } from 'react-router-dom';
+import { render } from '@testing-library/react';
+import configureStore from 'redux-mock-store';
+import { mockUser } from 'mocks/mockUser';
+import { SnackbarProvider } from 'notistack';
+import { mockCollection } from 'mocks/mockCollection';
+import { advanceTo, clear } from 'jest-date-mock';
+import SiteMetrics from '.';
+
+const mockStore = configureStore([]);
+
+describe('Monitoring SiteMetrics Page', () => {
+  beforeAll(() => {
+    advanceTo(new Date('2023-11-23T12:00:00'));
+  });
+
+  afterAll(() => {
+    clear();
+  });
+
+  let element: HTMLElement;
+  beforeEach(() => {
+    const store = mockStore({
+      user: {
+        userInfo: mockUser,
+        loading: false,
+        error: null,
+      },
+      collection: {
+        details: mockCollection,
+        loading: false,
+        error: null,
+      },
+    });
+
+    store.dispatch = jest.fn();
+
+    element = render(
+      <SnackbarProvider>
+        <ReduxProvider store={store}>
+          <BrowserRouter>
+            <SiteMetrics />
+          </BrowserRouter>
+        </ReduxProvider>
+      </SnackbarProvider>,
+    ).container;
+  });
+
+  it('should render with given state from Redux store', () => {
+    expect(element).toMatchSnapshot();
+  });
+});
